fix(rescue): extract fast-keepalive/respawn module under 1000-line cap + tokio Instant for paused-time test
BLOCKER 1 (file-size CI gate): endpoint_task.rs was 1255 lines (over the
1000-line *.rs cap). Pure-move extractions into new sibling modules bring it
to 899 lines, all new files well under 1000:
- endpoint_traits.rs    — ChunkFetcher/OutputProcess/OutputProcessFactory traits
                          + PrefetchedChunk (re-exported at endpoint_task level)
- fast_keepalive_escalation.rs — KeepaliveOutcome + keepalive_until_chunk (C1 #251)
- endpoint_handle.rs    — EndpointHandle struct + lifecycle methods
- endpoint_respawn.rs   — C3 (#237) producer-respawn decision helper
All moved items are re-exported so every existing crate::endpoint_task::* import
path (and the tests' super::super::super::* paths) keep resolving unchanged.
No logic change.

BLOCKER 2 (test CI gate): rust_rescue_push.rs used std::time::Instant in the
refill counter, so under #[tokio::test(start_paused = true)]
(rescue_push_resumes_normal_when_producer_recovers) it read the real wall clock
and continuous_active_secs never reached RESCUE_REFILL_TARGET_SECS (120) — the
loop never exited and the test panicked. Switch to tokio::time::Instant (honors
tokio::time::advance() under paused time; tracks wall time in prod). Duration
unchanged. Production behaviour unchanged.
